feat: put toJSON in its own lib function
